// generated/src/db/Enums.gen.ts
/* TypeScript file generated from Enums.res by genType. */

/* eslint-disable */
/* tslint:disable */

export type ContractType_t = "AcalEscrow";

export type EntityType_t = 
    "GlobalStats"
  | "Order"
  | "OrderEvent"
  | "dynamic_contract_registry";

export type OrderEventType_t = 
    "CREATED"
  | "LOCKED"
  | "COMPLETED"
  | "CANCELLED"
  | "DISPUTED";

export type OrderStatus_t = 
    "OPEN"
  | "LOCKED"
  | "COMPLETED"
  | "CANCELLED"
  | "DISPUTED"
  | "EXPIRED";